fix:flush topic json to disk

+++ src/main/java/com/sharkchili/minimq/broker/model/CommitLog.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class CommitLog {
     private String fileName;
-    private Long offset;
-    private Long limit;
+    private int offset;
+    private int limit;
 }
